+++ src/collective/behavior/amp/utils.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+"""Utility to transform HTML into AMP HTML.
+
+See: https://www.ampproject.org/docs/reference/spec.html
+"""
 from collective.behavior.amp.config import AMP_INVALID_ELEMENTS
 from collective.behavior.amp.logger import logger
 from lxml import html
@@ -58,6 +62,25 @@ class Html2Amp:
             parent.remove(tag)
             logger.debug('<{0}> tag was removed'.format(tag.tag))
 
+    def remove_invalid_attributes(self, el):
+        """Remove AMP HTML invalid attributes.
+        :param el: [required] LXML element to be sanitized.
+        :type el: lxml.html.HtmlElement
+        """
+        for tag in el.iterdescendants():
+            for key in tag.attrib.keys():
+                remove = False
+                if key.startswith('on') and key != 'on':
+                    remove = True
+                elif key == 'style':
+                    remove = True
+                elif key == 'xmlns' or key.startswith('xml:'):
+                    remove = True
+                if remove:
+                    del tag.attrib[key]
+                    logger.debug(
+                        '"{0}" attribute was removed from tag <{1}>'.format(key, tag.tag))
+
     def __call__(self, code):
         el = html.fromstring(code)
 
@@ -69,4 +92,5 @@ class Html2Amp:
 
         self.transform_img_tags(el)
         self.remove_invalid_tags(el)
+        self.remove_invalid_attributes(el)
         return html.tostring(el)