+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a3 (unreleased)
 ------------------
 
+- Clean up invalid AMP HTML attributes (fixes `#26`_).
+  [rodfersou]
+
 - Explicitly disable any Diazo theme on the AMP HTML view.
   [hvelarde]
 
@@ -27,3 +30,5 @@ There's a frood who really knows where his towel is.
 ------------------
 
 - Initial release.
+
+.. _`#26`: https://github.com/collective/collective.behavior.amp/issues/26